@@ -1,28 +1,24 @@
 'use strict'
-const gracefulFs = require('graceful-fs')
+const defaultFS = require('fs')
 const pathTemp = require('path-temp')
 
-module.exports = (dir, customFs) => {
-  const fs = customFs || gracefulFs
-  return new Promise((resolve, reject) => {
-    const tempFile = pathTemp(dir)
-    fs.writeFile(tempFile, '', 'utf8', err => {
-      if (!err) {
-        fs.unlink(tempFile, () => {})
-        resolve(true)
-        return
-      }
-      if (err.code === 'EACCES' || err.code === 'EPERM' || err.code === 'EROFS') {
-        resolve(false)
-        return
-      }
-      reject(err)
-    })
-  })
+module.exports = async (dir, customFS) => {
+  const fs = customFS || defaultFS
+  const tempFile = pathTemp(dir)
+  try {
+    await fs.promises.writeFile(tempFile, '', 'utf8')
+    fs.promises.unlink(tempFile).catch(() => {})
+    return true
+  } catch (err) {
+    if (err.code === 'EACCES' || err.code === 'EPERM' || err.code === 'EROFS') {
+      return false
+    }
+    throw err
+  }
 }
 
-module.exports.sync = (dir, customFs) => {
-  const fs = customFs || gracefulFs
+module.exports.sync = (dir, customFS) => {
+  const fs = customFS || defaultFS
   const tempFile = pathTemp(dir)
   try {
     fs.writeFileSync(tempFile, '', 'utf8')
