@@ -1,18 +1,15 @@
 'use strict'
-const fs = require('graceful-fs')
+const fs = require('fs').promises
 const path = require('path')
 const { promisify } = require('util')
 const rimraf = promisify(require('rimraf'))
 
-const mkdir = promisify(fs.mkdir)
-const readdir = promisify(fs.readdir)
-
 module.exports = async function makeEmptyDir (dir, opts) {
   if (opts && opts.recursive) {
-    await mkdir(path.dirname(dir), { recursive: true })
+    await fs.mkdir(path.dirname(dir), { recursive: true })
   }
   try {
-    await mkdir(dir)
+    await fs.mkdir(dir)
     return 'created'
   } catch (err) {
     if (err.code === 'EEXIST') {
@@ -24,7 +21,7 @@ module.exports = async function makeEmptyDir (dir, opts) {
 }
 
 async function removeContentsOfDir (dir) {
-  const items = await readdir(dir)
+  const items = await fs.readdir(dir)
   for (const item of items) {
     await rimraf(path.join(dir, item))
   }
