@@ -1,32 +1,29 @@
 'use strict'
-const fs = require('graceful-fs')
+const fs = require('fs')
 const path = require('path')
 const promisify = require('util').promisify
 const rimraf = promisify(require('rimraf'))
 const rimrafSync = require('rimraf').sync
 
-const rename = promisify(fs.rename)
-const mkdir = promisify(fs.mkdir)
-
 module.exports = async function renameOverwrite (oldPath, newPath) {
   try {
-    await rename(oldPath, newPath)
+    await fs.promises.rename(oldPath, newPath)
   } catch (err) {
     switch (err.code) {
       case 'ENOTEMPTY':
       case 'EEXIST':
         await rimraf(newPath)
-        await rename(oldPath, newPath)
+        await fs.promises.rename(oldPath, newPath)
         break
       // weird Windows stuff
       case 'EPERM':
         await new Promise(resolve => setTimeout(resolve, 200))
         await rimraf(newPath)
-        await rename(oldPath, newPath)
+        await fs.promises.rename(oldPath, newPath)
         break
       case 'ENOENT':
-        await mkdir(path.dirname(newPath), { recursive: true })
-        await rename(oldPath, newPath)
+        await fs.promises.mkdir(path.dirname(newPath), { recursive: true })
+        await fs.promises.rename(oldPath, newPath)
         break
       default:
         throw err
