@@ -1,11 +1,8 @@
 'use strict'
 const path = require('path')
-const fs = require('graceful-fs')
+const fs = require('fs')
 const writeFileAtomic = require('write-file-atomic')
 const YAML = require('js-yaml')
-const { promisify } = require('util')
-
-const mkdir = promisify(fs.mkdir)
 
 const main = (fn, fp, data, opts) => {
   if (!fp) {
@@ -24,7 +21,7 @@ const main = (fn, fp, data, opts) => {
 }
 
 module.exports = async (fp, data, opts) => {
-  await mkdir(path.dirname(fp), { recursive: true })
+  await fs.promises.mkdir(path.dirname(fp), { recursive: true })
   return main(writeFileAtomic, fp, data, opts)
 }
 
